requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pyswisseph==2.10.3.2
pytz==2025.1
timezonefinder==6.5.2
geopy==2.4.1
